Vastly improve html/CSS for image grid
Most changes come from the decision to re-evaluate the resizing behavior
of the image layout. The old design was overcomplicated and often
unpredictable, and often achieved its layout by resizing the images
themselves. The new approach was based on rearranging images instead of
resizing them.

Restructured html removes extraneous div tag (class="box"). The confusing
position properties of the containers and images were changed to
inline-box. The resulting layout is coherent, and has the advantage of
being fully dynamic with no reliance on media queries.

Some cleanup remains. js code that deals with aspect ratios is likely
useless and should be deleted.