Move image loading to deferred framework, introduce titleObjects
The two ajax-based events required to fetch a usable article and get its
images are now expressed neatly in a single line via jQuery's Deferred
utility. This clearly exposes the code's execution flow and relieves the
hassle of following chained callbacks when debugging.

I've also gotten a few functions closer to implementing guessing on
titles. Written a constructor for a titleObject, designed to store/parse page
titles and guesses and keep records of which words have been successfully
guessed.