Fix input bar behavior and begin introducing footer
Restored input bar functionality after introducing a few bugs in the last
update. Changes to the html introduce a new site footer element that will
contain site information at the bottom of the page. A few new lines of js
introduce scroll and resize listeners that monitor vertical position on
the page and prevent the fixed-display input bar at the bottom of the
sreen from overlapping it.