Makes tons of changes to UI implementation
Lots and lots of HTML/CSS changes in this commit. I was already finding
the html difficult to manage, so I rewrote the footer widget from
scratch. The result is much better code. I removed a truly staggering
amount of unnecessary html and css, and reconstructed the thing with good
practice in mind. This might not even be the last rewrite I end up doing,
seeing as I'm still going into a lot of things blind. Learning!