Make fixes to style and behavior of footer and answer widget
Changes:
-Fix box shadows on widget (hide widget header shadows when menu is not
extended)
-Relocate footer style from skeleton.scss to its own sass partial,
_pagefooter.scss
-Remove page footer in mobile layout due to inconsistent widget
positioning in mobile browsers. Relocate it to the share menu, share.html,
where it uses the same styling as did on the main page. It looks really
bad and should be updated. Also, it may be a good idea to change the menu
button from 'share' to something like 'about' to reflect this change.
-Show mobile menu every time a new history entry is submitted.
-Fix desktop scrolling when history log extends beyond mainrow. This was
causing a problem when the user attempted to scroll beyond the bottom of the
game panel, because the menu position would switch to fixed, causing the
page to resize. The fix involves adding an invisible spacer of the same
height as the menu that switches to position: relative at the same time
the menu becomes fixed.