Implement win/lose behavior
Winning, by inputting the correct article title, or losing, by clicking
the 'give up' button, now triggers an overlay with the article title, a
link to the page, and, if the player won, their final score (number of
guesses, images, and hints). The overlay also presents a "play again"
button that refreshes the page, and a "share" button which is not yet
implemented.

Additional changes:
-Fixed clipping on panel shadows in desktop layout
-Menu shadows become lighter and more diffuse when the position changes
from relative to fixed
-Counter elements now incremented via a general incrementCounter function.