Straighten out some clip overlap behavior, reduce lag on mobile
Floating widget is now drawn on its own layer thanks to the transform: translateZ(0); css hack. This is especially important for reducing lag when the mobile keyboard is opened.

The page footer is back at the bottom of the mobile layout. It also includes space after the footer for the expanded widget so that the last page image can be viewed even when the menu is open.