Implement more robust url parsing, replace sharing buttons.
Also roll back change regarding loading thumbnails instead of full-size images on mobile. This didn't seem to be making a huge difference in scrolling speed, and on a solid connection, the delay incurred as the server generated the thumbnails made image time significantly greater than when full-size images were used.